requirements.txt
django
django-allauth
django-cors-headers
django-rest-auth
djangorestframework
gunicorn
whitenoise
django-widget-tweaks
drf-yasg
coreapi
django-cleanup
gunicorn
django-heroku
boto3
Pillow
django-storages
stripe
requests
python-decouple
django-countries